Removed all references of helper functions from src/algorithms/index.js that only retrieved metadata (strings about algorithms) to now referene the pure masterList. Also removed all these helper functions from src/algorithms/index.js and into src/algorithms/masterList.js
