+++ src/components/AlgorithmMenu.js
@@ -4,12 +4,12 @@ import "../styles/AlgorithmMenu.scss";
 // Get the base URL dynamically
 const baseUrl = window.location.origin;
   
-import { DeployedAlgorithms } from '../algorithms/masterList';
+import { DeployedAlgorithmCategoryList } from '../algorithms/masterList';
 
 // Dynamically create URLs in algorithm menu from master list (src/algorithms/index.js).
 // Only entries with noDeploy=false are included.
 const algorithms = Object.fromEntries(
-  DeployedAlgorithms.map(({ category, algorithms }) => [
+  DeployedAlgorithmCategoryList.map(({ category, algorithms }) => [
     category,
     Object.fromEntries(
       algorithms.map(({ name, shorthand, mode }) => [
