+++ src/context/GlobalState.js
@@ -4,7 +4,8 @@ import React, { createContext, useState } from 'react';
 import { initGlobalAlgorithmGetter } from '../algorithms/controllers/collapseChunkPlugin';
 import { initGlobalAlgotithmGetter } from '../algorithms/controllers/transitiveClosureCollapseChunkPlugin';
 import { dispatcher, initialState } from './actions';
-import algorithms, { getDefaultMode, getCategory } from '../algorithms';
+import algorithms from '../algorithms';
+import { getDefaultMode, getCategory } from '../algorithms/masterList'
 import { Search } from '@mui/icons-material';
 import { URLProvider, URLContext } from './urlState';
